UI first cut

+++ src/main/java/com/vsr/dao/DepartmentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface DepartmentRepository extends PagingAndSortingRepository<Department, Long> {
+public interface DepartmentRepository extends PagingAndSortingRepository<Department, Integer> {
 }
