+++ src/main/java/com/vsr/dao/FaremapRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface FaremapRepository extends PagingAndSortingRepository<Faremap, Long> {
+public interface FaremapRepository extends PagingAndSortingRepository<Faremap, Integer> {
 }
