+++ src/main/java/com/vsr/dao/StationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface StationRepository extends PagingAndSortingRepository<Station, Long> {
+public interface StationRepository extends PagingAndSortingRepository<Station, Integer> {
 }
